+++ packages/plugin_platform_interface/lib/plugin_platform_interface.dart
@@ -29,7 +29,7 @@ import 'package:meta/meta.dart';
 ///   /// Platform-specific plugins should set this with their own platform-specific
 ///   /// class that extends [UrlLauncherPlatform] when they register themselves.
 ///   static set instance(UrlLauncherPlatform instance) {
-///     PlatformInterface.verifyExtends(instance, _token);
+///     PlatformInterface.verify(instance, _token);
 ///     _instance = instance;
 ///   }
 ///
@@ -58,22 +58,22 @@ abstract class PlatformInterface {
   ///
   /// This is implemented as a static method so that it cannot be overridden
   /// with `noSuchMethod`.
-  static void verifyExtends(PlatformInterface instance, Object token) {
+  static void verify(PlatformInterface instance, Object token) {
     if (identical(instance._instanceToken, const Object())) {
       throw AssertionError('`const Object()` cannot be used as `token`.');
     }
-    _verifyExtends(instance, token);
+    _verify(instance, token);
   }
 
-  /// Performs the same checks as `verifyExtends` but without throwing an
+  /// Performs the same checks as `verify` but without throwing an
   /// [AssertionError] if `const Object()` is used as the instance token.
   ///
   /// This method will be deprecated in a future release.
   static void verifyToken(PlatformInterface instance, Object token) {
-    _verifyExtends(instance, token);
+    _verify(instance, token);
   }
 
-  static void _verifyExtends(PlatformInterface instance, Object token) {
+  static void _verify(PlatformInterface instance, Object token) {
     if (instance is MockPlatformInterfaceMixin) {
       bool assertionsEnabled = false;
       assert(() {